User permission task

+++ src/pages/api/teams/[teamId]/bots/[botId]/index.js
@@ -6,6 +6,7 @@ import { bentoTrack } from '@/lib/bento'
 import { mpTrack } from '@/lib/mixpanel'
 import { deleteBot } from '@/lib/apiFunctions'
 import { validateBotParams } from '@/lib/apiFunctions'
+import { canUserEditBot, canUserCreateDeleteBot } from '@/utils/function.utils'
 
 export default async function handler(req, res) {
   configureFirebaseApp()
@@ -21,6 +22,12 @@ export default async function handler(req, res) {
   const { botId } = req.query
 
   if (req.method === 'PUT') {
+    //check user is allowed to edit bot or not
+    if (!canUserEditBot(team, userId)) {
+      return res.status(402).json({
+        message: 'You are not allowed to edit this bot.',
+      })
+    }
     try {
       const bot = await getBot(team.id, botId)
       if (!bot) {
@@ -55,6 +62,12 @@ export default async function handler(req, res) {
       return res.status(500).json({ message: error?.message })
     }
   } else if (req.method === 'DELETE') {
+    //check user is allowed to delete bot or not
+    if (!canUserCreateDeleteBot(team, userId)) {
+      return res.status(402).json({
+        message: 'You are not allowed to delete bot.',
+      })
+    }
     try {
       await deleteBot(team.id, botId)
